requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==24.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-speech==2.30.0
google-cloud-texttospeech==2.23.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langsmith==0.2.10
MarkupSafe==3.0.2
marshmallow==3.25.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.59.6
orjson==3.10.14
packaging==24.2
pip==24.2
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==75.1.0
simple-websocket==1.1.0
sniffio==1.3.1
SpeechRecognition==3.13.0
SQLAlchemy==2.0.37
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.3.0
uv==0.5.16
Werkzeug==3.1.3
wheel==0.44.0
wsproto==1.2.0
yarl==1.18.3
